feat(boards): give the async agent live search/research tools while authoring
Follow-up: the document author now has on-demand retrieval instead of relying
only on the graph's pre-fetched context.

- Extract createSearchTools (gruenerator_search, web_search, research, examples,
  pressemitteilung_examples, + optional direct_response) from the legacy
  chatStreamController into a shared agents/searchTools.ts. The chat handler keeps
  its router behaviour via { includeDirectResponse: true }; the board agent omits
  it and runs tools on-demand.
- boardAgentWorker passes these tools to generateText with stopWhen
  stepCountIs(5) (toolChoice auto), so the model can search/research before
  writing. Prompt updated to instruct active tool use.
